ROME-Python/requirements.txt
numba
numpy
pyserial
pyfirmata
spatialmath-python
# DOCS https://petercorke.github.io/spatialmath-python/
# https://petercorke.github.io/robotics-toolbox-python/index.html